moved backend functionality to service
